add post request

+++ main.go
@@ -2,43 +2,54 @@ package main
 
 import (
 	"fmt"
+	"log"
 	"net/http"
 	"os/exec"
 )
 
-func getVmStatus(w http.ResponseWriter, r *http.Request) {
-	cmd := exec.Command("bash", "/Virsh-Api-Server/getVmStatus.sh")
-	output, _ := cmd.Output()
-	fmt.Fprintf(w, string(output))
-	fmt.Println("getVmStatus")
+func getVmStatus(w http.ResponseWriter, r *http.Request) { // POST
+	if r.Method == "POST" {
+		cmd := exec.Command("bash", "/Virsh-Api-Server/getVmStatus.sh")
+		output, _ := cmd.Output()
+		fmt.Fprintf(w, string(output))
+		log.Println("getVmStatus")
+	}
 }
 
-func startVM(w http.ResponseWriter, r *http.Request) {
-	cmd := exec.Command("sudo", "virsh", "start", "Windows10")
-	output, _ := cmd.Output()
-	fmt.Fprintf(w, string(output))
-	fmt.Println("startVM")
+func startVM(w http.ResponseWriter, r *http.Request) { // POST
+	if r.Method == "POST" {
+		cmd := exec.Command("sudo", "virsh", "start", "Windows10")
+		output, _ := cmd.Output()
+		fmt.Fprintf(w, string(output))
+		log.Println("startVM")
+	}
 }
 
-func shutdownVM(w http.ResponseWriter, r *http.Request) {
-	cmd := exec.Command("sudo", "virsh", "shutdown", "Windows10")
-	output, _ := cmd.Output()
-	fmt.Fprintf(w, string(output))
-	fmt.Println("shutdownVM")
+func shutdownVM(w http.ResponseWriter, r *http.Request) { // POST
+	if r.Method == "POST" {
+		cmd := exec.Command("sudo", "virsh", "shutdown", "Windows10")
+		output, _ := cmd.Output()
+		fmt.Fprintf(w, string(output))
+		log.Println("shutdownVM")
+	}
 }
 
-func forceShutdownVM(w http.ResponseWriter, r *http.Request) {
-	cmd := exec.Command("sudo", "virsh", "destroy", "Windows10")
-	output, _ := cmd.Output()
-	fmt.Fprintf(w, string(output))
-	fmt.Println("forceShutdownVM")
+func forceShutdownVM(w http.ResponseWriter, r *http.Request) { // POST
+	if r.Method == "POST" {
+		cmd := exec.Command("sudo", "virsh", "destroy", "Windows10")
+		output, _ := cmd.Output()
+		fmt.Fprintf(w, string(output))
+		log.Println("forceShutdownVM")
+	}
 }
 
-func rebootVM(w http.ResponseWriter, r *http.Request) {
-	cmd := exec.Command("sudo", "virsh", "reboot", "Windows10")
-	output, _ := cmd.Output()
-	fmt.Fprintf(w, string(output))
-	fmt.Println("rebootVM")
+func rebootVM(w http.ResponseWriter, r *http.Request) { // POST
+	if r.Method == "POST" {
+		cmd := exec.Command("sudo", "virsh", "reboot", "Windows10")
+		output, _ := cmd.Output()
+		fmt.Fprintf(w, string(output))
+		log.Println("rebootVM")
+	}
 }
 
 func handleRequests() {
